Ballistic Analysis of the Hypotheses

Look at the overall characteristics of the hypotheses, such as number of regression models and how they will be put together based on the approach. Is used to evaluate the objects created by bullet() and by aim() (which also uses ballistics to generate the appropriate modeling matrix).

ballistics(mark, ...)

# S3 method for default
ballistics(mark, ...)

# S3 method for bullet
ballistics(mark, ...)

# S3 method for aim
ballistics(mark, ...)

Arguments

mark
Represents either a bullet or aim object to be analyzed for further information or characteristics

A single bullet object will return a table that is structured for the proposed analytical approach

A single aim object, after the fire() function, will return a table that includes further analytical elements. Currently returns the runtime in miliseconds.

...
For extensibility

Value

The function returns values based on what object was given to it.

A bullet object returns a tibble with the following columns:

outcomes: character vector

model_num: an identifying model number

formulas: a list of formulas that will be used for modeling

model_spec: the parsnip model specification

An aim object can be returned in two different ways.

If it has not yet been processed, it returns a table with additional columns identifying potential problems before it can be processed by the fire() function

If it has been processed by the fire() function, it will return additional columns to the original table that include model statistics, including a tidied fitted object

Examples

library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() %>% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "parallel", model = lm_mod)
ballistics(b)

#> # A tibble: 6 x 5
#>   outcomes model_num predictors formulas  model_spec
#>   <chr>        <int> <list>     <list>    <list>    
#> 1 mpg              1 <chr [1]>  <formula> <spec[+]> 
#> 2 mpg              2 <chr [2]>  <formula> <spec[+]> 
#> 3 mpg              3 <chr [2]>  <formula> <spec[+]> 
#> 4 cyl              1 <chr [1]>  <formula> <spec[+]> 
#> 5 cyl              2 <chr [2]>  <formula> <spec[+]> 
#> 6 cyl              3 <chr [2]>  <formula> <spec[+]>