Create bullets to be loaded

A bullet object is a pre-specified hypothesis, with a formula that helps to identify the individual variables and their roles. It requires the use of a model specification from parsnip (or any other similarly specified models). It also determines what sequence of models to put together.

bullet(f, exposure = NULL, approach = "sequential", model, ...)

Arguments

f
Formula showing relationship of outcomes and predictors

exposure
Variable(s) that is forced to be maintained in every model as a predictor.

approach
The modeling approach that will be used. The options are:

sequential will build y ~ x1, y ~ x1 + x2 models

parallel will build y ~ x1, y ~ x2 models

model
A model specified by the parsnip package, which includes the mode and computational engine

...
For extensibility

Value

Object of type bullet that contains the hypothesis and the characteristics of how to put the analysis together. needed for analysis

Examples

library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() %>% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "sequential", model = lm_mod)